refactor: add country table row component
