+++ src/components/shared/CountryTableRow/country.table.row.module.scss
@@ -1,0 +1,19 @@
+.checkbox {
+  vertical-align: middle;
+}
+
+.name {
+  display: flex;
+  align-items: center;
+}
+
+.col {
+  line-height: 1.5;
+}
+
+.flag {
+  max-width: 3rem;
+  height: auto;
+
+  margin-right: 1rem;
+}
