+++ src/components/shared/TableCountries/index.tsx
@@ -1,24 +1,13 @@
 import cn from 'classnames';
 import React from 'react';
 
-import {
-  Select,
-  Table,
-  TableBody,
-  TableCell,
-  TableCellHead,
-  TableContainer,
-  TableHead,
-  TableRow,
-} from '@/components/ui';
+import { CountryAutocomplete, CountryTableRow, NoResults, Preloader } from '@/components/shared';
+import { Select, Table, TableBody, TableCellHead, TableContainer, TableHead, TableRow } from '@/components/ui';
 import { useCountryFilters } from '@/hooks';
 import { useFetchCountriesQuery } from '@/services/api';
-import { useAppDispatch, useAppSelector } from '@/store';
-import { addVisited, removeVisited } from '@/store/visited/slice';
 import { Regions } from '@/types';
-import { createRegions, formatByRanks } from '@/utils';
+import { createRegions } from '@/utils';
 
-import { CheckboxVisited, CountryAutocomplete, NoResults, Preloader } from '../';
 import classes from './table.countries.module.scss';
 
 interface TableProps {
@@ -35,8 +24,6 @@ export function TableCountries({ className }: Readonly<TableProps>) {
   const [sortField, setSortField] = React.useState<SortField>('population');
   const [sortOrder, setSortOrder] = React.useState<SortOrder>('desc');
   const filteredCountries = useCountryFilters(data, selectedRegion, searchQuery, sortField, sortOrder);
-  const visitedCountries = useAppSelector((state) => state.visited.countries);
-  const dispatch = useAppDispatch();
 
   if (isFetching) {
     return <Preloader />;
@@ -99,50 +86,8 @@ export function TableCountries({ className }: Readonly<TableProps>) {
                 </TableRow>
               </TableHead>
               <TableBody>
-                {filteredCountries?.map((country, idx) => (
-                  <TableRow
-                    key={country.cca2}
-                    onClick={() => {
-                      dispatch(addVisited(country));
-                    }}
-                  >
-                    <TableCell className={classes.col} onClick={(e) => e.stopPropagation()}>
-                      <CheckboxVisited
-                        className={classes.checkbox}
-                        isVisited={visitedCountries.some((c) => c.cca2 === country.cca2)}
-                        onVisited={() => {
-                          const isVisited = visitedCountries.some((c) => c.cca2 === country.cca2);
-                          if (isVisited) {
-                            dispatch(removeVisited(country.cca2));
-                          } else {
-                            dispatch(addVisited(country));
-                          }
-                        }}
-                      />
-                    </TableCell>
-                    <TableCell align="center" className={classes.col}>
-                      {idx + 1}
-                    </TableCell>
-                    <TableCell className={classes.col} align="left">
-                      <div className={classes.name}>
-                        {country.flags.png && (
-                          <img className={classes.flag} src={country.flags.png} alt={country.name.common} />
-                        )}
-                        {country.name.common}
-                      </div>
-                    </TableCell>
-                    <TableCell className={classes.col}>{country.region}</TableCell>
-                    <TableCell className={classes.col}>{country.capital}</TableCell>
-                    <TableCell className={classes.col}>{formatByRanks(country.area)}</TableCell>
-                    <TableCell className={classes.col}>{formatByRanks(country.population)}</TableCell>
-                    <TableCell className={classes.col}>
-                      {Object.entries(country.currencies).map(([code, currency]) => (
-                        <div key={code}>
-                          {currency.name} ({currency.symbol})
-                        </div>
-                      ))}
-                    </TableCell>
-                  </TableRow>
+                {filteredCountries.map((country, idx) => (
+                  <CountryTableRow key={country.cca2} country={country} idx={idx} />
                 ))}
               </TableBody>
             </Table>
