+++ src/components/shared/TableCountries/table.countries.module.scss
@@ -1,29 +1,9 @@
 .root {
 }
 
-.flag {
-  max-width: 3rem;
-  height: auto;
-
-  margin-right: 1rem;
-}
-
-.name {
-  display: flex;
-  align-items: center;
-}
-
-.col {
-  line-height: 1.5;
-}
-
 .controls {
   display: flex;
   max-width: 60%;
   gap: 2rem;
   margin-bottom: 3.5rem;
-}
-
-.checkbox {
-  vertical-align: middle;
 }
